+++ lib/yito/model/booking/ysd_md_booking_category.rb
@@ -34,6 +34,7 @@ module Yito
         property :active, Boolean, :default => true
         property :web_public, Boolean, :default => true
         property :alias, String, :length => 80
+        property :color, String, length: 50
 
         property :from_price, Decimal, :scale => 2, :precision => 10, :default => 0
         property :from_price_offer, Decimal, :scale => 2, :precision => 10, :default => 0