+++ lib/yito/model/booking/ysd_md_booking_queries_mysql.rb
@@ -45,59 +45,6 @@ module Yito
           BookingDataSystem::Booking.by_sql{ |b| [first_customer_booking_query(b,params)] }.all(order: [:creation_date.desc], offset:0, limit: 1).first
         end
 
-        # Incoming money summary
-        #
-        def incoming_money_summary(year)
-
-          query = <<-QUERY
-             SELECT DATE_FORMAT(date_from, '%Y-%m') as period, 
-                 sum(total_cost) as total
-             FROM bookds_bookings
-             WHERE status IN (2,3,4) and YEAR(date_from) = #{year.to_i}
-             GROUP BY period
-             ORDER by period
-          QUERY
-
-          summary = @repository.adapter.select(query)
-
-        end
-
-        #
-        # Get the reservations received grouped by month
-        #
-        def reservations_received(year)
-       
-          query = <<-QUERY
-             SELECT DATE_FORMAT(creation_date, '%Y-%m') as period, 
-                 count(*) as occurrences
-             FROM bookds_bookings
-             WHERE YEAR(creation_date) = #{year.to_i}
-             GROUP BY period
-             ORDER by period
-          QUERY
-
-          reservations=@repository.adapter.select(query)
-
-        end
-
-        #
-        # Get the reservations confirmed grouped by month
-        #
-        def reservations_confirmed(year)
-       
-          query = <<-QUERY
-             SELECT DATE_FORMAT(creation_date, '%Y-%m') as period, 
-                 count(*) as occurrences
-             FROM bookds_bookings
-             WHERE status NOT IN (1,5) and YEAR(creation_date) = #{year.to_i}
-             GROUP BY period 
-             ORDER by period
-         QUERY
-
-         reservations=@repository.adapter.select(query)
-
-        end
-
         # Received reservations
         #
         def count_received_reservations(year)