+++ lib/yito/model/booking/ysd_md_booking_queries_postgresql.rb
@@ -65,51 +65,6 @@ module Yito
           @repository.adapter.select(query).first
         end
 
-        def incoming_money_summary(year)
-
-          query = <<-QUERY
-             SELECT TO_CHAR(date_from, 'YYYY-MM') as period, 
-                 sum(total_cost) as total
-             FROM bookds_bookings
-             WHERE status NOT IN (1,5) and date_part('year', date_from) = #{year.to_i}
-             GROUP BY period
-             ORDER by period
-          QUERY
-
-          summary = @repository.adapter.select(query)
-
-        end
-
-        def reservations_received(year)
-       
-          query = <<-QUERY
-             SELECT TO_CHAR(creation_date, 'YYYY-MM') as period, 
-                 count(*) as occurrences
-             FROM bookds_bookings 
-             WHERE date_part('year', creation_date) = #{year.to_i}
-             GROUP BY period
-             order by period
-          QUERY
-
-          reservations=@repository.adapter.select(query)
-
-        end
-
-        def reservations_confirmed(year)
-       
-          query = <<-QUERY
-             SELECT TO_CHAR(creation_date, 'YYYY-MM') as period, 
-                  count(*) as occurrences
-             FROM bookds_bookings
-             WHERE status NOT IN (1,5) and date_part('year', creation_date) = #{year.to_i}
-             GROUP BY period 
-             order by period
-          QUERY
-
-          reservations=@repository.adapter.select(query)
-
-        end
-
         def count_received_reservations(year)
  
           query = <<-QUERY