+++ lib/yito/model/booking/ysd_md_booking_queries_sqlite.rb
@@ -45,49 +45,6 @@ module Yito
           BookingDataSystem::Booking.by_sql{ |b| [first_customer_booking_query(b,params)] }.all(order: [:creation_date.desc], offset:0, limit: 1).first
         end
 
-        def incoming_money_summary(year)
-
-          query = <<-QUERY
-             SELECT TO_CHAR(date_from, 'YYYY-MM') as period, 
-                 sum(total_cost) as total
-             FROM bookds_bookings
-             WHERE status IN (2,3,4) and strftime('%Y', date_from) = #{year.to_i}
-             GROUP BY period
-             ORDER by period
-          QUERY
-
-          summary = @repository.adapter.select(query)
-
-        end
-
-        def reservations_received(year)
-       
-          query = <<-QUERY
-             SELECT TO_CHAR(creation_date, 'YYYY-MM') as period, 
-                 count(*) as occurrences
-             FROM bookds_bookings and strftime('%Y', creation_date) = #{year.to_i}
-             GROUP BY period
-             order by period
-          QUERY
-
-          reservations=@repository.adapter.select(query)
-
-        end
-
-        def reservations_confirmed(year)
-       
-          query = <<-QUERY
-             SELECT TO_CHAR(creation_date, 'YYYY-MM') as period, 
-                  count(*) as occurrences
-             FROM bookds_bookings
-             WHERE status NOT IN (1,5) and strftime('%Y', creation_date) = #{year.to_i}
-             GROUP BY period 
-             order by period
-          QUERY
-
-          reservations=@repository.adapter.select(query)
-
-        end
 
         def count_received_reservations(year)